Move bbox computation to base object
